requirements.txt
pyyaml
pandas
numpy
great-expectations
google-cloud-bigquery
google-cloud-storage
psycopg2-binary
mysql-connector-python
google-auth
google-auth-oauthlib
google-auth==2.3.0
sqlalchemy-bigquery